backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
google-generativeai
